requirements.txt
asgiref==3.2.10
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-extensions==3.0.9
django-filter==2.4.0
django-mathfilters==1.0.0
django-phonenumber-field==5.0.0
et-xmlfile==1.0.1
install==1.3.4
jdcal==1.4.1
numpy==1.19.4
openpyxl==3.0.5
phonenumbers==8.12.10
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
